% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-add_model.R
\name{add_mechanistic}
\alias{add_mechanistic}
\title{Add a mechanistic model to a dadnow or multidadnow object}
\usage{
add_mechanistic(
  dadnow,
  formula,
  params = list(sc = 0.2, sp = 0.3, method = "poisson")
)
}
\arguments{
\item{dadnow}{A dadnow or multidadnow object.}

\item{Dt, Ct, Pt, Rt}{The data correspoding to DAD, CNISP, PTSOS, and RVDSS, respectively.}

\item{Rt_nowcast}{The RVDSS data for the nowcast period. This is used to create the nowcast predictions for the mechanistic model.}

\item{sc, sp}{The scaling factors for the CNISP and PTSOS data.}

\item{method}{Either "normal", "poisson", or "negbinom".}
}
\value{
A dadnow or multidadnow object with the mechanistic model added.
}
\description{
Add a mechanistic model to a dadnow or multidadnow object
}
