% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-add_model.R
\name{add_model.multidadnow}
\alias{add_model.multidadnow}
\title{Add a model to an existing multidadnow object}
\usage{
\method{add_model}{multidadnow}(multidadnow, formula = NULL, model, params = NULL)
}
\arguments{
\item{multidadnow}{A multidadnow object.}

\item{formula}{A formula object.}

\item{model}{The model to use for nowcasting. Currently implemented: "lm", "ar". Can be a vector, in which case the model is trained for each model in the vector.}

\item{params}{The parameters to use for the model. Must be a named list.}
}
\value{
A multidadnow object with the model added.
}
\description{
Add a model to an existing multidadnow object
}
