% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-autoplot.R
\name{autoplot.dadnow}
\alias{autoplot.dadnow}
\title{Plot the nowcasts}
\usage{
\method{autoplot}{dadnow}(dadnow, last_n = NULL)
}
\description{
Deprecated - "dadnow" objects are no longer used, but I don't want to break anything.
}
