% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prep_data.R
\name{check_all_NAs}
\alias{check_all_NAs}
\title{Check if any of the data columns are all NAs (all missing)}
\usage{
check_all_NAs(data)
}
\arguments{
\item{data}{Dataframe of data as expected by the function \code{nowcast()}.}
}
\value{
An error message
}
\description{
Check if any of the data columns are all NAs (all missing)
}
\keyword{internal}
