% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-enbpi.R
\name{enbpi}
\alias{enbpi}
\title{Ensemble batch prediction intervals and prediction metrics}
\usage{
enbpi(
  X_train,
  y_train,
  model,
  formula,
  params,
  k,
  batches = 40,
  train_window = NULL,
  se = TRUE
)
}
\arguments{
\item{X_train}{A data frame of training data.}

\item{y_train}{A vector of training data.}

\item{model}{The model to use for nowcasting.}

\item{formula}{A formula object.}

\item{params}{The parameters to use for the model.}

\item{k}{The number of steps ahead to predict. If NULL, the number of steps is determined by the amount of data to be nowcast.}

\item{batches}{The number of batches to use for training (akin to the number of folds for k-fold cross validation).}

\item{train_window}{The number of days to use for training. Defaults to 60\% of the training data, which allows for a large training set for each batch while also allowing for a reasonable amount of variation in the test sets.}
}
\value{
A list containing the k-step ahead prediction standard errors and the evaluations of the models for the k-step ahead predictions.
}
\description{
EnbPI takes \code{batches} of training data, fits a model to each batch, and then uses the residuals to calculate prediction intervals and prediction metrics. This targets the exact standard errors that we're looking for. If we're doing 8-step ahead nowcasting, we need an estimate of the standard error for the eighth step ahead, not just the general standard error for predictions. This is a model-agnostic approach to calculating prediction intervals.
}
