% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prep_data.R
\name{find_nas}
\alias{find_nas}
\title{Return the number of NA's at the end of a vector (internal)}
\usage{
find_nas(y)
}
\arguments{
\item{y}{A vector of values with some number of NA's at the end of it.}
}
\description{
Return the number of NA's at the end of a vector (internal)
}
\keyword{internal}
