% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ar.R
\name{fit_ar}
\alias{fit_ar}
\title{Fit an AR model on given data and make predictions for a given set of data with options for using an AR, ARX, or ARIMA model}
\usage{
fit_ar(Y_train, X_train = NULL, X_nowcast = NULL, params = list(p = 1))
}
\arguments{
\item{params}{A named list containing additional parameters: \code{p} Integer indicating the degree of the AR model, and \code{n.ahead} integer indicating the number of predictions to be make}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Current data of X_train for which there is no Y_train data, used to make nowcasts}
}
\value{
Arima object, predictions, and fitted values of the model on the training data
}
\description{
Fit an AR model on given data and make predictions for a given set of data with options for using an AR, ARX, or ARIMA model
}
