% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-basicModels.R
\name{fit_basic_models}
\alias{fit_basic_models}
\title{Fit a basic indicated model}
\usage{
fit_basic_models(formula, model, data, ar_order = 1)
}
\arguments{
\item{formula}{A formula indicating which variables to use in the model}

\item{model}{A string indicating the type of model to be fit to the data, can be linear or AR}

\item{data}{A data frame containing at least the data to be used}

\item{ar_order}{An optional argument indicating the order of the AR model to be used}
}
\value{
The specified model based on the given formula
}
\description{
Fit a basic indicated model
}
\examples{
y <- rnorm(100)
x <- seq(1,100)
df <- data.frame(x,y)
fit-basicModels(y ~ x, model = "linear", df)
}
