% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GAM.R
\name{fit_GAM}
\alias{fit_GAM}
\title{Fit a spline model on given data and make predictions for a given set of data}
\usage{
fit_GAM(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  family = gaussian,
  smooths = list()
)
}
\arguments{
\item{family}{Model family, as used in \code{mgcv::gam()}. Defaults to \code{gaussian} and accepts any family that is accepted by \code{mgcv::gam()}.}

\item{smooths}{Named list of smoothing parameters for each variable, e.g. \code{list("x1" = 10, "x2")} will result in a formula \code{s(x1, bs = "tp", k = 10) + s(x2, bs = "tp", k = -1) + ...}. If a smoothing parameter is not specified, it will default to -1, which will result \code{mgcv::gam()} choosing the smoothing parameter.}

\item{XTrain}{Training data for the explanatory variables in the model}

\item{YTrain}{Training data for the response variable}

\item{XNowcast}{Data to make predictions bases on}
}
\value{
GAM model object and predictions
}
\description{
Fit a spline model on given data and make predictions for a given set of data
}
