% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_KalmanFilter.R
\name{fit_KalmanFilter}
\alias{fit_KalmanFilter}
\title{Fit a Kalman Filter model on given data and make predictions for a given set of data}
\usage{
fit_KalmanFilter(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(degree = 1)
)
}
\arguments{
\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Data to make predictions bases on3}

\item{degree}{For trend component, integer defining the degree of the polynomial trend}

\item{CovMatrix}{Covariance matrix or array of disturbance terms $\epsilon_t$ of observation equation}
}
\value{
Kalman Filter model object and predictions
}
\description{
Fit a Kalman Filter model on given data and make predictions for a given set of data
}
