% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanistic.R
\name{fit_mechanistic}
\alias{fit_mechanistic}
\title{Fit a mechanistic model to the data}
\usage{
fit_mechanistic(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(sc = 0.2, sp = 0.3, method = "normal")
)
}
\arguments{
\item{X_train, Y_train, X_nowcast}{The data correspoding to DAD, CNISP, PTSOS, and RVDSS, respectively.}

\item{params}{The parameters to use for the model. Must be a named list containing sc and sp and method (normal, poisson, or negbinom).}
}
\value{
A list with the parameter estimates and the negative log-likelihood.
}
\description{
Fit a mechanistic model to the data
}
\examples{
sim_poisson_data <- function(eta, Rt, sc = 0.2, sp = 0.3) {
  data.frame(
    Dt = rpois(length(Rt), eta * Rt),
    Ct = rpois(length(Rt), eta * sc * Rt),
    Pt = rpois(length(Rt), eta * sp * Rt),
    Rt = Rt
  )
}
d2 <- sim_poisson_data(eta = 5, Rt = rpois(100, 100))
fit_mechanistic(Y_train = d2$Dt, X_train = d2["Ct", "Pt", "Rt"], X_nowcast = rpois(10, 100), params = list(method = "poisson"))
}
