% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RF.R
\name{fit_RF}
\alias{fit_RF}
\title{Fit a Random Forest model on given data and make predictions for a given set of data}
\usage{
fit_RF(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  ntree = 500,
  mtry = if (!is.null(Y_train) && !is.factor(Y_train)) {
     max(floor(ncol(X_train)/3),
    1)
 } else {
     floor(sqrt(ncol(X_train)))
 },
  weights = NULL,
  replace = TRUE,
  maxnodes = NULL
)
}
\arguments{
\item{ntree}{Integer indicating the number of trees in the random forest, default is 500}

\item{mtry}{Indicates the number of features to try in each node of the Random Forest. Default is the default for randomForest}

\item{weights}{A vector of length same as y that are positive weights used only in sampling data to grow each tree (not used in any other calculation)}

\item{replace}{Should sampling of cases be done with or without replacement?}

\item{maxnodes}{Maximum number of terminal nodes trees in the forest can have.}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Data to make predictions bases on}
}
\value{
Random Forest object and predictions
}
\description{
Fit a Random Forest model on given data and make predictions for a given set of data
}
