% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-basicModels.R
\name{fit_simple_arx}
\alias{fit_simple_arx}
\title{Fit a simple ARX model (internal)}
\usage{
fit_simple_arx(x, y, order = 1)
}
\arguments{
\item{x}{The single covariate vector.}

\item{y}{The response vector.}

\item{order}{The order of the AR(p) process.}
}
\description{
Fit a simple ARX model (internal)
}
