% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-basicModels.R
\name{fit_simple_lm}
\alias{fit_simple_lm}
\title{Fit a simple linear model (internal)}
\usage{
fit_simple_lm(x, y)
}
\arguments{
\item{x}{The single covariate vector.}

\item{y}{The response vector.}
}
\description{
Fit a simple linear model (internal)
}
