% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-basicModels.R
\name{fit_simple_model}
\alias{fit_simple_model}
\title{Fit a simple (on covariate) regression model}
\usage{
fit_simple_model(formula, data, model = c("lm", "ar")[1], order = NULL)
}
\arguments{
\item{formula}{y ~ x}

\item{data}{A data frame with columns used in the formula.}

\item{model}{Currently implemented: "lm" and "ar"}

\item{order}{The AR order to be used in \code{arima()}.}
}
\value{
A model object for which a predict method exists.
}
\description{
Fit a simple (on covariate) regression model
}
