% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_XGBoost.R
\name{fit_XGBoost}
\alias{fit_XGBoost}
\title{Fit an eXtreme Gradient Boost model model on given data and make predictions for a given set of data}
\usage{
fit_XGBoost(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(nrounds = 100, verbose = 1, XGBparams = list())
)
}
\arguments{
\item{params}{A named list containing options for the parameters used by \code{xgboost::xgb.train}}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Data to make predictions bases on}
}
\value{
XGBoost object and predictions
}
\description{
Fit an eXtreme Gradient Boost model model on given data and make predictions for a given set of data
}
