% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-get_attr.R
\name{get_formula}
\alias{get_formula}
\title{Get the formula used in the underlying model from a dadnow object}
\usage{
get_formula(dadnow, model_id)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{model_id}{The model id to use for the model.}
}
\value{
The underlying model, in it's original form.
}
\description{
Get the formula used in the underlying model from a dadnow object
}
