% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-get_attr.R
\name{get_predictions}
\alias{get_predictions}
\title{Get the predictions from a dadnow object}
\usage{
get_predictions(dadnow, model_ids = NULL)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{model_ids}{A model name or vector of model names. Must correspond to the model ids seen in the evaluations table. Note that the model names can change depending on the other models added to the dadnow object, so always double check.}
}
\value{
A matrix of predictions, with rows as dates and columns as models.
}
\description{
Get the predictions from a dadnow object
}
