% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-impute.R
\name{impute_linear}
\alias{impute_linear}
\title{Basic linear interpolation imputation}
\usage{
impute_linear(dates, x)
}
\arguments{
\item{dates}{A vector of dates.}

\item{x}{A vector of values.}
}
\value{
A vector of values with NAs imputed.
}
\description{
Basic linear interpolation imputation
}
