% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now_simple_models.R
\name{now_simple_lm}
\alias{now_simple_lm}
\title{Nowcast from a simple linear model (internal)}
\usage{
now_simple_lm(newX, model)
}
\arguments{
\item{model}{The model to use for predictions.}

\item{NewX}{The known x values to be nowcast on.}
}
\description{
Nowcast from a simple linear model (internal)
}
