% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now_simple_models.R
\name{now_simple_models}
\alias{now_simple_models}
\title{Perform nowcasting for the specified response variable}
\usage{
now_simple_models(formula, data, model = c("lm", "ar")[1], order = NULL)
}
\arguments{
\item{formula}{y ~ x}

\item{data}{A data frame with columns used in the formula.}

\item{model}{Currently implemented: "lm" and "ar"}

\item{order}{The AR order to be used in \code{arima()}.}
}
\value{
Predictions for the missing data in the response variable
}
\description{
Perform nowcasting for the specified response variable
}
