% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcast from a dadnow object}
\usage{
nowcast(dadnow, quiet = FALSE)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{quiet}{If TRUE, suppresses output.}
}
\value{
A dadnow object with the nowcasts added.
}
\description{
Nowcast from a dadnow object
}
