% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-nowcast_one.R
\name{nowcast_mechanistic}
\alias{nowcast_mechanistic}
\title{Fit a mechanistic model to the data, returning a dadnow object}
\usage{
nowcast_mechanistic(
  formula,
  data,
  batches = 40,
  train_window = NULL,
  level = 0.95,
  date_col = NULL,
  params = list(sc = 0.2, sp = 0.3, method = "normal")
)
}
\arguments{
\item{formula}{A formula object, \emph{must} be of the form Dt ~ Ct + Pt + Rt.}

\item{data}{A data frame. Must contain the variables specified in the formula and in \code{date_col}. Trailing NA values in \code{y} will be nowcasted.}

\item{date_col}{Name of the column containing date information. If NULL, the date information attempted to be inferred. If there's a single datetime column then it is used. If the data are a ts or mts or zoo object, the dates are esxtracted.}

\item{params}{The parameters to use for the model. Must be a named list containing sc and sp and method (normal, poisson, or negbinom).}
}
\value{
A dadnow object with the mechanistic model added.
}
\description{
Fit a mechanistic model to the data, returning a dadnow object
}
