% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-nowcast2.R
\name{nowcast_one}
\alias{nowcast_one}
\title{Nowcast various models from a prepared data frame}
\usage{
nowcast_one(formula, data, model, params, date_col = NULL)
}
\arguments{
\item{formula}{A formula, e.g. y ~ x, y ~ lag(x1, 1) + lag(x2, 3)}

\item{data}{A data frame. Must contain the variables specified in the formula and in \code{date_col}. Trailing NA values in \code{y} will be nowcasted.}

\item{model}{The model to use for nowcasting. Currently implemented: "lm", "ar". Can be a vector, in which case the model is trained for each model in the vector.}

\item{params}{The parameters to use for the model. Must be a named list.}

\item{date_col}{Name of the column containing date information. If NULL, the date information attempted to be inferred. If there's a single datetime column then it is used. If the data are a ts or mts or zoo object, the dates are esxtracted.}
}
\value{
An object of class "\code{dadnow}".
}
\description{
Nowcast various models from a prepared data frame
}
