% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-methods-dadnow.R
\name{print.multidadnow}
\alias{print.multidadnow}
\title{Print a dadnow object}
\usage{
\method{print}{multidadnow}(dadnow)
}
\arguments{
\item{dadnow}{A dadnow object.}
}
\value{
A dadnow object invisibly.
}
\description{
Print a dadnow object
}
