% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulated matrix with reasonable columns}
\usage{
simulate_data(
  cnisp_samp_prop = 0.2,
  ptsos_samp_prop = 0.3,
  virulence = 0.05,
  to_nowcast = 10
)
}
\description{
Simulation for testing/example purposes
}
