% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-methods-dadnow.R
\name{summary.dadnow}
\alias{summary.dadnow}
\title{Summarize a dadnow object}
\usage{
\method{summary}{dadnow}(dadnow)
}
\arguments{
\item{dadnow}{A dadnow object.}
}
\value{
A dadnow object invisibly.
}
\description{
Currently a placeholder.
}
