% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-train_models.R
\name{train_model}
\alias{train_model}
\title{Train a model for nowcasting}
\usage{
train_model(dadnow, quiet = FALSE)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{quiet}{If TRUE, suppresses output.}
}
\value{
A dadnow object with the model trained.
}
\description{
Train a model for nowcasting
}
