% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-add_model.R
\name{add_model}
\alias{add_model}
\title{Add a model to an existing dadnow or multidadnow object}
\usage{
add_model(x, ...)
}
\arguments{
\item{x}{A dadnow or multidadnow object.}

\item{formula}{A formula object.}

\item{model}{The model to use for nowcasting. Currently implemented: "lm", "ar". Can be a vector, in which case the model is trained for each model in the vector.}

\item{params}{The parameters to use for the model. Must be a named list.}
}
\value{
A dadnow or multidadnow object with the model added.
}
\description{
Add a model to an existing dadnow or multidadnow object
}
