% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-autoplot.R
\name{autoplot.multidadnow}
\alias{autoplot.multidadnow}
\title{Plot the nowcasts}
\usage{
\method{autoplot}{multidadnow}(multidadnow, last_n = NULL)
}
\arguments{
\item{multidadnow}{A multidadnow object.}

\item{last_n}{The number of most recent observations to plot (in addition to the nowcasts). Defaults to all.}
}
\description{
Plot the nowcasts
}
