% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-enbpi.R
\name{enbpi}
\alias{enbpi}
\title{Ensemble batch prediction intervals and prediction metrics}
\usage{
enbpi(
  X_train,
  y_train,
  model,
  formula,
  params,
  k,
  batches = 40,
  train_window = NULL,
  level = 0.95
)
}
\arguments{
\item{X_train}{A data frame of training data.}

\item{y_train}{A vector of training data.}

\item{model}{The model to use for nowcasting.}

\item{formula}{A formula object.}

\item{params}{The parameters to use for the model.}

\item{k}{The number of steps ahead to predict. If NULL, the number of steps is determined by the amount of data to be nowcast.}

\item{batches}{The number of batches to use for training (akin to the number of folds for k-fold cross validation).}

\item{train_window}{The number of days to use for training. Defaults to 60\% of the training data, which allows for a large training set for each batch while also allowing for a reasonable amount of variation in the test sets.}

\item{level}{The prediction interval level.}
}
\value{
A dadnow object with the ensemble predictions and intervals added.
}
\description{
Ensemble batch prediction intervals and prediction metrics
}
