% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-enbpi.R
\name{enbpi}
\alias{enbpi}
\title{Ensemble batch prediction intervals and prediction metrics}
\usage{
enbpi(dadnow, model, k = NULL, train_window = 30, level = 0.95)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{k}{The number of steps ahead to predict.}

\item{train_window}{The number of days to use for training.}

\item{level}{The prediction interval level.}

\item{batches}{The number of batches to use for training.}
}
\value{
A dadnow object with the ensemble predictions and intervals added.
}
\description{
Ensemble batch prediction intervals and prediction metrics
}
