% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_AR.R
\name{fit_AR}
\alias{fit_AR}
\title{Fit an AR model on given data and make predictions for a given set of data with options for using an AR or ARX model}
\usage{
fit_AR(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  p = 1,
  d = 0,
  q = 0,
  n.ahead = 1
)
}
\arguments{
\item{p}{Integer indicating the number of parameters for the AR model}

\item{n.ahead}{Integer indicating the number of predictions to be make}

\item{XTrain}{Training data for the explanatory variables in the model}

\item{YTrain}{Training data for the response variable}

\item{XNowcast}{Data to make predictions bases on}
}
\value{
Linear model object and predictions
}
\description{
Fit an AR model on given data and make predictions for a given set of data with options for using an AR or ARX model
}
