% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_arx.R
\name{fit_arx}
\alias{fit_arx}
\title{Fit an ARX model on given data and make predictions for a given set of data}
\usage{
fit_arx(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(p = 1, d = 0, q = 0, n.ahead = 1)
)
}
\arguments{
\item{params}{Named list with additional parameters: \code{p} Integer indicating the number of parameters for the AR model}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Current data of X_train for which there is no Y_train data, used to make nowcasts}
}
\value{
Arima model object, predictions, and fitted values of the model on training data
}
\description{
Fit an ARX model on given data and make predictions for a given set of data
}
