% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LM.R
\name{fit_LM}
\alias{fit_LM}
\title{Fit a linear model on given data and make predictions for a given set of data}
\usage{
fit_LM(Y_train, X_train, X_nowcast, params = NULL)
}
\arguments{
\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Data to make predictions bases on}
}
\value{
Linear model object and predictions
}
\description{
Fit a linear model on given data and make predictions for a given set of data
}
