% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcast various models from a prepared data frame}
\usage{
nowcast(
  formula,
  data,
  model,
  params = NULL,
  date_col = NULL,
  batches = 40,
  train_window = NULL,
  level = 0.95,
  se = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{formula}{A formula, e.g. y ~ x, y ~ lag(x1, 1) + lag(x2, 3)}

\item{data}{A data frame. Must contain the variables specified in the formula and in \code{date_col}. Trailing NA values in \code{y} will be nowcasted.}

\item{model}{The model to use for nowcasting. See \code{vignette("Model_Details", package = "dadnowcast")} and \code{vignette("Mechanistic_Model", package = "dadnowcast")} for more information about currently implemented models, and \code{vignette("Write_fit_XX_functions", package = "dadnowcast")} for how to write your own model for use in this package.}

\item{params}{The parameters to use for the model. Must be a named list.}

\item{date_col}{Name of the column containing date information. If NULL, the date information attempted to be inferred. If there's a single datetime column then it is used. If the data are a ts or mts or zoo object, the dates are esxtracted.}

\item{batches}{The number of batches to use for training in the EnbPI calculation (akin to the number of folds for k-fold cross validation).}

\item{train_window}{The number of days to use for training. Defaults to 60\% of the training data, which allows for a large training set for each batch while also allowing for a reasonable amount of variation in the test sets.}

\item{level}{The prediction interval level.}

\item{se}{Should the standard errors and evaluation metrics be calculated by ensemble-based batch prediction intervals (EnbPI)? If FALSE, the function will return NAs for the standard errors and evaluation metrics. Can also be set to "theoretical" to use the theoretical prediction intervals for the model, if they are returned by the fit_*() function.}

\item{quiet}{Logical. If TRUE, silences all output messages.}
}
\value{
An object of class "\code{dadnow}".
}
\description{
Nowcast various models from a prepared data frame
}
\details{
A short overview of the specific models used here is given below, more details can be found in the vignette \code{Model_Details} and the individual help files (\verb{?fit_*}, where \code{*} is the name of the model).
\itemize{
\item "lm", a standard linear model, which takes no additional parameters.
\item "ar", an autoregressive model, it has the parameter \code{p} which is the degree of the AR model.
\item "arx", an autoregressive model with exogenous variables, it is an AR model that adds the capacity to consider explanatory variables, it has the same parameters as an ar model.
\item "arima", an ARIMA model combines an AR model with a moving average (MA) model, it has parameters \code{p}, which specifies the degree of the AR component, \code{d} which specifies the degree of differencing in the model, and \code{q} which is the degree of the moving average model.
\item "gam", a generalized additive model, it has parameters \code{family} to specify the family used, \code{smooths} which is smoothing parameters for each variable, and \code{bs} which specifies basis smoothing functions.
\item "kf", a Kalman filter model, which is a continuously updating model with a prediction and update portion, it takes additional parameters \code{degree} which specifies the degree of the trend, and \code{CovMatrix} which is a covariance matrix for the disturbance terms.
\item "rf", a random forest model, a tree based model that grows many trees and combines the output to pick a prediction, it takes parameters \code{ntree} the number of trees to grow, \code{mtry} the number of variables used as candidates at each split, \code{weights} weights the sample data, \code{replace} should sampling be done with or without replacement, \code{maxnodes} limits the number of terminal nodes, \code{nodesize} sets the minimum size of the terminal nodes.
\item "xgboost", a extreme gradient boosting model, which is a tree based model in which subsequent trees learn from previous trees, it takes parameters \code{nrounds} which is the number of boosting iteration to do, and \code{XGBparams} a list of more parameters used in the model the most important of which is \code{max_depth} both \code{nrounds} and \code{max_depth} are tuned if their values are not specified, it also takes \code{verbose} should output be silent (0) or not (1).
\item "mechanistic", creates a mechanistic model, \code{sc}, \code{sp}, and \code{method} specifies the family.
\item See \code{vignette("Write_fit_XX_functions", package = "dadnowcast")} for how to write your own model for use in this package.
}
}
