% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare the data for analysis, returning an object ready for further analysis steps.}
\usage{
prep_data(
  formula,
  data,
  model,
  date_col = NULL,
  interpolate = TRUE,
  folds = 5,
  cross_val_indices = NULL,
  quiet = TRUE
)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A data frame.}

\item{model}{The model to use for nowcasting. Currently implemented: "lm", "ar". Can be a vector, in which case the model is trained for each model in the vector.}

\item{date_col}{Name of the column containing date information. If NULL, the date information attempted to be inferred. If there's a single datetime column then it is used. If the data are a ts or mts or zoo object, the dates are esxtracted.}

\item{interpolate}{Whether to interpolate missing values. Defaults to TRUE.}

\item{folds}{(Deprecated) The number of folds to use for cross validation. Defaults to 5.}

\item{cross_val_indices}{(Deprecated) The indices of the cross validation folds. Defaults to NULL.}

\item{quiet}{Whether to print messages. Defaults to TRUE.}
}
\value{
Object of class dadnow
}
\description{
Prepare the data for analysis, returning an object ready for further analysis steps.
}
\keyword{internal}
