% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-simulate_data.R
\name{simulate_mechanistic}
\alias{simulate_mechanistic}
\title{Simulate from the mechanistic model}
\usage{
simulate_mechanistic(
  to_nowcast = 10,
  ar1 = 0.65,
  ar2 = 0.3,
  sigma = 0.1,
  eta = 0.2,
  sc = 0.2,
  sp = 0.3
)
}
\arguments{
\item{to_nowcast}{The number of days to nowcast.}

\item{ar1, ar2, sigma}{The AR parameters for the RVDSS.}

\item{eta}{The scaling factor for the CNISP and PTSOS and DAD.}

\item{sc, sp}{The scaling factors for the CNISP and PTSOS data.}
}
\value{
A dataframe with simulated data for DAD, CNISP, PTSOS, and RVDSS.
}
\description{
Simulate from the mechanistic model
}
